+++ src/main/java/com/example/demo/repository/goals/GoalsRepository.java
@@ -1,4 +1,4 @@
-package com.example.demo.dao.goals;
+package com.example.demo.repository.goals;
 
 import com.example.demo.entity.Goals;
 import org.springframework.data.repository.CrudRepository;
