+++ src/main/java/com/example/demo/repository/tasks/TasksDAO.java
@@ -1,4 +1,4 @@
-package com.example.demo.dao.tasks;
+package com.example.demo.repository.tasks;
 
 import com.example.demo.entity.Tasks;
